Nearly eliminated Arduino dependency
1. Use char* and std::string instead of Arduino String
2. Use m5 APIs for buttons, millis, etc
3. Move most system dependencies into System.cpp

This is a bit step toward making FluidDial run on Windows.

The one remaining Arduino dependency is on the LittleFS class, for the
purpose of loading PNG files, and that dependency is isolated in
System.cpp.  The Windows version of the m5 framework can load PNG
files via C standard IO filesystem calls, so the LittleFS dependency
will go away for the Windows build.